Move update loops to loops/
